Fix 13 issues: concurrency, context threading, transactions, resilience
- Use read transaction for estimateInterval in runReportDevice (#232)
- Fix Logout auth race: send request before clearing logged flag (#233)
- Thread parent context through connectClient (#234)
- Give each connectClient retry its own 10s timeout (#235)
- Deep-copy big.Int and byte slices in requestSnapshot (#236)
- Wrap chat context DB queries in read-only transaction (#237)
- Add 60s idle timeout for Ollama streaming (#238)
- Use atomicWriteFile in runSetup (#239)
- Migrate WAL/SHM files alongside database (#240)
- Fix integer division in formatConnBreakdown percentages (#241)
- Add cancellable context to dashboard for clean quit (#242)
- Save partial chat response on streaming error (#243)
- Show router-unreachable message in diffNetworkSnapshots (#244)

+++ output.go
@@ -406,8 +406,8 @@ func formatConnBreakdown(breakdown map[string]int64) string {
 
 	var parts []string
 	for _, k := range keys {
-		pct := breakdown[k] * 100 / total
-		parts = append(parts, fmt.Sprintf("%s:%d%%", connAbbrev(k), pct))
+		pct := float64(breakdown[k]) * 100.0 / float64(total)
+		parts = append(parts, fmt.Sprintf("%s:%.0f%%", connAbbrev(k), pct))
 	}
 	return strings.Join(parts, " ")
 }
